Move builder from src to test
