+++ test/builder/asset-gatherer.ts
