+++ test/builder/tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "compileOnSave": true,
+    "compilerOptions": {
+      "lib": [
+        "es2017",
+        "webworker"
+      ],
+      "sourceMap": false,
+      "moduleResolution": "node",
+      "target": "esnext",
+      "module": "esnext",
+      "allowJs": false,
+      "noUnusedLocals": true,
+      "strictNullChecks": true,
+      "noImplicitAny": true,
+      "noImplicitThis": true,
+      "alwaysStrict": true
+    },
+    "include": [
+      "*.ts",
+      "**/*.ts"
+    ]
+  }
